+++ pixelfed/CHANGELOG.md
@@ -1,3 +1,10 @@
+0.6
+
+* Version bump for new base image 14.1
+* Extra steps to trim image size
+
+---
+
 0.5
 
 * Version bump for new base image